Modified UI and added new State functionality: - Added Weight increments - Added logic to round to 0.1 digits

+++ composeApp/src/commonMain/kotlin/org/jdcoding/workouttracker/workoutZone/domain/Workout.kt
@@ -6,7 +6,7 @@ data class Workout(
     val sets: Int = 0,
     val reps: Int = 0,
     val duration: String? = "",
-    val weight: Float?,
+    val weight: Double?,
     val weightUnit: String?,
     val icon: String?
 )
